Backend/requirements.txt
alembic==1.18.4
aniso8601==10.0.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
Mako==1.3.10
MarkupSafe==3.0.3
pycparser==3.0
python-dotenv==1.2.2
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
WTForms==3.2.1
